requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
qdrant-client==1.11.0
sentence-transformers==3.1.0
pydantic==2.9.0
httpx==0.27.0
requests==2.32.3
jinja2==3.1.4
duckduckgo-search==6.2.0
beautifulsoup4==4.12.3
lxml==5.3.0

